import from utils/ims-file-path to separate node-type modules from browser-safe ones for import trimming; update tsconfig placement to accommodate
